fix(web): use server-side authentication instead of middleware
