+++ apps/web/src/app/api/auth/logout/route.ts
@@ -2,15 +2,15 @@ import { hashToken } from "@/lib/auth";
 import { getDB } from "@/lib/db";
 import { refreshTokens } from "@/lib/db/schema/refresh-tokens";
 import { eq } from "drizzle-orm";
-import { NextResponse } from "next/server";
+import { NextRequest, NextResponse } from "next/server";
 import z from "zod";
 
 const logoutFormat = z.object({
   refreshToken: z.string(),
 });
 
-export async function POST(response: NextResponse) {
-  const data = await response.json();
+export async function POST(request: NextRequest) {
+  const data = await request.json();
   const validationResult = logoutFormat.safeParse(data);
   if (!validationResult.success) {
     return NextResponse.json(
