+++ apps/web/src/app/api/me/route.ts
@@ -1,3 +1,4 @@
+import { ensureAuthenticated } from "@/lib/auth";
 import { getDB } from "@/lib/db";
 import { accounts } from "@/lib/db/schema/accounts";
 import { users } from "@/lib/db/schema/users";
@@ -5,7 +6,11 @@ import { eq } from "drizzle-orm";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(request: NextRequest) {
-  const accountID = request.headers.get("x-account-id")!;
+  const authResponse = ensureAuthenticated(request);
+  if (!authResponse.success) {
+    return authResponse.failResponse!;
+  }
+  const accountID = authResponse.accountID!;
   const user = await getDB()
     .select()
     .from(accounts)
